# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server",
    "version": "1.0.0",
    "title": "Swagger TFG"
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    },"/register": {
      "post": {
        "tags": [
          "register"
        ],
        "summary": "Register",
        "description": "",
        "operationId": "register",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Register parameters",
            "required": true,
            "schema": {
              "$ref": "#/definitions/register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/loginSuccess"
            }
          },
          "500": {
            "description": "operation fail",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        },
        "security": [
          {
            "authorization": [
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "authorization": {
      "type": "Authorization",
      "name": "Bearer",
      "in": "header"
    }
  },
  "definitions": {
...
    "loginSuccess": {
      "type":"object",
      "properties": {
        "accessToken": {
          "type":"string"
        },
        "tokenType": {
          "type": "string",
          "enum": ["Bearer" , "Token"]
        }
      }
    },
    "register": {
      "type": "object",
      "properties": {
        "nick": {
          "type": "string"
        },
        "password": {
          "type": "string",
          "format": "hash"
        },
        "name": "string",
        "surname": "string"
      }
    },
...
    "error": {
      "type": "object",
      "properties": {
        "status": {
          "type": "number"
        },
		    "message": {
          "type": "string"
        }
      }
    }
  }
}